more instructions

+++ src/examples/MyClass.java
@@ -56,7 +56,7 @@ public class MyClass {
 		}
 	}
 	
-	public Node myMethod2(float i, float j) {
+	public Node myMethod2(double i, double j) {
 		i = i + j;
 		if (i - 1 > 0) {
 			i = 2 * i - 1;
@@ -75,8 +75,8 @@ public class MyClass {
 //		Node x = new Node();
 //		Node y = new Node();
 //		mc.myMethod();
-		float i = 0;
-		float j = 0;
+		double i = 0;
+		double j = 0;
 		mc.myMethod2(i, j);
 //		mc.myMethod1(x, y);
 	}
